+++ client/src/admin/components/Delete.js
